+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.6 (2021-01-13)
+
+## Chore
+
+- @capacitor-community/sqlite@2.4.5-3
+
+### Bug Fixes
+
+- remove android permision issue#60
+- fix README & CHANGELOG
+
 ## 0.0.5 (2020-12-29)
 
 ## Chore